Fix water overload and safety warning alerts
- Add safety warning evaluation to session end process
- Connect SafetyWarningDialog to display alerts when appropriate
- Fix type conversion issues in safety calculations
- Safety alerts now properly appear and dismiss like kidney overload alerts
- Resolves missing safety dialog triggers

+++ lib/services/hydration_calculator.dart
@@ -168,7 +168,7 @@ class HydrationCalculator {
 
   /// Evaluate session safety and return appropriate warning
   SafetyWarning? evaluateSessionSafety(double ml, Duration duration) {
-    final ratePerHour = (ml / duration.inMinutes) * 60;
+    final ratePerHour = (ml / duration.inMinutes.toDouble()) * 60;
     
     // Danger level - immediate health risk
     if (ml > 500 && duration.inMinutes < 10) {
